Rename file: `chain-id-error` -> `errors`
