+++ libs/protocol-client/src/index.ts
@@ -1,4 +1,4 @@
 export * from './lib/protocol-client';
 export * from './lib/protocol-types';
-export * from './lib/utils/chain-id-error';
+export * from './lib/utils/errors';
 export { default } from './lib/utils/patch';
