+++ libs/protocol-client/src/lib/utils/errors.ts
